Nueva BD, Arreglo de Pedidos, modelos de pgs PagePay y PageShipping

+++ Dominio/Pedido.cs
@@ -10,7 +10,7 @@ namespace Dominio
     public class Pedido
     {
 
-        public int IdPedido { get; set; }
+        public long IdPedido { get; set; }
         public string NombreUsuario { get; set; }
         public string Email { get; set; }
         public DateTime Fecha { get; set; }
